move math module to utils
