+++ cubic/shared/src/main/scala/org/hammerlab/math/FromDouble.scala
