+++ cubic/shared/src/test/scala/cubic/complex/CubicTest.scala
@@ -1,6 +1,6 @@
 package cubic.complex
 
-import hammerlab.math.FromDouble
+import org.hammerlab.math.FromDouble
 import org.hammerlab.math.polynomial.TestCase
 import org.hammerlab.math.polynomial.test.PolySolverTest
 import org.hammerlab.math.syntax.Doubleish
