+++ format/shared/src/main/scala/hammerlab/math/sigfigs.scala
@@ -1,0 +1,9 @@
+package hammerlab.math
+
+import org.hammerlab.math.format.HasSigFigs
+
+trait sigfigs
+  extends HasSigFigs
+    with Serializable
+
+object sigfigs extends sigfigs
