--- math/js/src/main/scala/hammerlab/math.scala
@@ -1,12 +1,0 @@
-package hammerlab
-
-import org.hammerlab.math.all
-
-/**
- * Expose various features for importing via:
- *
- * {{{
- * import hammerlab.math._
- * }}}
- */
-object math extends all
