+++ utils/js/src/main/scala/hammerlab/math/utils.scala
@@ -1,0 +1,5 @@
+package hammerlab.math
+
+import org.hammerlab.math.all
+
+object utils extends all
