+++ utils/jvm/src/main/scala/hammerlab/math/utils.scala
@@ -1,0 +1,17 @@
+package hammerlab.math
+
+import org.hammerlab.math.all
+
+/**
+ * Expose various features for importing via:
+ *
+ * {{{
+ * import hammerlab.math._
+ * }}}
+ */
+object utils
+  extends all {
+  import org.hammerlab.{ math ⇒ ohm }
+  type HypergeometricDistribution = ohm.HypergeometricDistribution
+  val  HypergeometricDistribution = ohm.HypergeometricDistribution
+}
