+++ utils/jvm/src/test/scala/org/hammerlab/math/VarNumTest.scala
