+++ utils/shared/src/main/scala/org/hammerlab/math/Binomial.scala
