+++ utils/shared/src/main/scala/org/hammerlab/math/Div.scala
