+++ utils/shared/src/main/scala/org/hammerlab/math/Interpolate.scala
