+++ utils/shared/src/main/scala/org/hammerlab/math/Min.scala
