+++ utils/shared/src/main/scala/org/hammerlab/math/VarNum.scala
