+++ utils/shared/src/main/scala/org/hammerlab/math/all.scala
@@ -1,0 +1,15 @@
+package org.hammerlab.math
+
+import org.hammerlab.{ math ⇒ ohm }
+
+trait all
+  extends HasBinomial
+     with Div
+     with Interpolate
+     with Min
+     with Steps {
+  type RoundNumbers[I] = ohm.RoundNumbers[I]
+  val  RoundNumbers = ohm.RoundNumbers
+
+  val Steps = ohm.Steps
+}
