+++ utils/shared/src/main/scala/org/hammerlab/math/show.scala
